added entity model

+++ AddressBook.Common/Models/AddressBookModel.cs
@@ -5,7 +5,7 @@
     public long Id { get; set; }
     public string Name { get; set; }
     public System.DateTime DateOfBirth { get; set; }
-    public int Contact { get; set; }
+    public long Contact { get; set; }
     public bool InTheCountry { get; set; }
   }
 }
